mcq_generation_using_genAI/requirements.txt
openai
langchain
python-dotenv
streamlit
PyPDF2

-e .
